requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
joblib
scikit-learn
psycopg2-binary